requirements_flask.txt
Flask==3.0.3
Werkzeug>=3.0
Jinja2>=3.1
itsdangerous>=2.1
click>=8.1

pydub>=0.25.1
python-docx>=1.1.2
openai>=1.46.0
httpx>=0.26.0

python-dotenv>=1.0.1
tqdm>=4.66
